Refactor database configuration and introduce JFR metrics tracking for HikariCP

+++ server/src/main/java/jeffrey/testapp/server/PersonController.java
@@ -78,7 +78,7 @@ public class PersonController {
 
     @GetMapping("/count")
     public ResponseEntity personCount() {
-        int count = repository.count();
+        int count = repository.count().intValue();
         LOG.info("Person Count: {}", count);
 
         if (efficientMode) {
